Heavily refactored Mesh class and texture loading
Mesh objects no longer need to always contain the highest amount of
available vertex data and can instead be customized as necessary. This
may introduce issues such as attempting to render an animatable object
as static when it has been loaded as a static mesh, but that should be
handled as an implementation issue at least for now.