Resource loading tweaks
ResourceConstants converted to an Enum with utility function for
building resource paths.
Resources are expected to be loaded from an archive always rather than
directly in the file system, thus requiring paths to be separated by a
forward slash rather than the OS separator.
Initial tweaks for individual resource loading changes to remove the
reliance on ResourceFileReader.