Separating event processing from update and render thread
Changed Avogine to start a separate thread for handling game updates and
rendering so that the application will no longer freeze if the event
polling is stalled like when the Window is dragged/resized.
Refactored Nuklear implementation to have its own render class separate
from the context.
Added plenty of TODOs mostly around input handling, it's likely that the
existing InputListeners will be largely toned down in favor of manual
input state checking in Game and its associated members.