Tearing up materials models and meshes
Once again changed the way VAO and VBO generation is handled.
In this would be experimental change, a Mesh is expected to implement a
VertexArrayObject, and the Mesh itself would represent the OpenGL VAO.
Materials are also updated to now contain the Meshes that they apply to,
this is largely how it already works but leads to slight confusion over
possible AnimatedMesh or InstancedMesh handling, which should be
offloaded to the consumer site knowing what types of Meshes it's
expecting.