Further changing jank VAO/VBO builder functions
To remove builder specific functions from the VAO and VBO record
instance fields, the VAO builder was modified to expose a new gen()
method that accepts an array of VBO Suppliers to allow for programmatic
creation and initialization of a VAO with its VBOs and attributes.
Its not quite as simple and sweet as the original implementation, but it
exists to prevent the possibility of trying to configure more VBOs in an
already constructed VAO or potentially modifying the vertex attributes
for a specific VBO element.